Current transformer reporting average transfer values with floating point precision

+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/BlockMetalDevices2.java
@@ -253,7 +253,7 @@ public class BlockMetalDevices2 extends BlockIEBase implements ICustomBoundingbo
 			if(!world.isRemote)
 			{
 				TileEntityEnergyMeter meter = (TileEntityEnergyMeter)te;
-				int transfer = meter.getAveragePower();
+				int transfer = (int)meter.getAveragePower();
 				if (meter.dummy) {
 					TileEntity tmp = world.getTileEntity(meter.xCoord, meter.yCoord+1, meter.zCoord);
 					if (tmp instanceof TileEntityEnergyMeter)
